P2 Director incl. EMF Compare Core only, intentionally not SDK w. UI
This makes it easier to switch from the (Mars) Target Platform included
in yangide (and still used by default, because that's more consistent)
to the "Running Platform" as Target, without getting compilation errors
in yangide ModelSynchronizer.

Switching TP to Running Platform is useful to verify differences between
Mars and e.g. Neon such as
https://bugs.eclipse.org/bugs/show_bug.cgi?id=492829 (nothing to do with
yangide).