Move Target Platform task into yangide.setup
This setup task used to take a fair amount of time, and is really only
required only ever for Eclipse plugin dev ... to save the 99% of users
time, it thus makes perfect sense to only do this as part of the yangide
dev env (not user) set up.

Now change activate of this task from "false" to "true".

Change-Id: I0a9c5454d5271cfe656cff7afb311da64fcfab31
Signed-off-by: Michael Vorburger <vorburger@redhat.com>