Maven offline, to avoid unexpected long running mvn JAR updates
in ODL this is a bit particular, compared to other projects; because
EVERYTHING is a SNAPSHOT, you typically end up waiting about once a day
for a very long time while ALL of ODL JARs are redownloaded.

I've come to the conclusion that it's best to let developers initiate
that explicitly, by a CLI mvn, or UI right-click Maven > Update Project,
then letting it "just happen", every day, as it currently does.

Change-Id: If7a740f93a2e0c7a538b0f1e9d0c3942275946b3
Signed-off-by: Michael Vorburger <vorburger@redhat.com>