Remove ${body_statement} from constructor body code template
Hoping (not properly tested and verified..) that this will prevent the
useless super() in default constructors which has been repeatedly
flagged during reviews of some of my changes; e.g- again today in
https://git.opendaylight.org/gerrit/#/c/63120/

According to the Eclipse Help this is for "The 'Constructor body'
templates are used when new method or constructor with body is created.
It contains the variable ${body_statement} that resolves a super call."
so hopefully removing this will do the trick.

Change-Id: I1572dfa2238635d76a4ba6a7192c873ae5d51814
Signed-off-by: Michael Vorburger <vorburger@redhat.com>